// src/app/views/trade/model/custAddress.model.ts
export class CustAddress {

  constructor() {}

  Cust_Code: string;
  Addr_Seq: string;
  Addr_No: string;
  Place: string;
  Road: string;
  Tambon_Id: string;
  Amphur_Id: string;
  Province_Id: string;
  Country_Id: string;
  Zip_Code: string;
  Print_Address: string;
  Tel: string;
  Fax: string;
}